Add TODO for Collider interface
A TODO to remind us to create a Collider interface so collision detection can be cleaner code.

Without a Collider interface we may end up checking for collisions with plants, then rocks, then wall, etc. With the interface we can just check all colliders. I think it will be cleaner.