Shrink gardener by 50%
Gardener now half the size it was before.

Also changes tile dimensions to 16x8 instead of 20x5

Also adjusted painting of watering can and gardener to compensate for grid size and empty space in the portion of their images extracted for sprites.

Also left a couple of debug rectangles visible - so I can see the rectangle of TILE_WIDTH x TILE_HEIGHT that corresponds to the location of the gardener and the watering can. These debug rectangles are what allowed me to make sure the sprites were properly placed relative to the objects position coords.