Play sound when gardener bumps
Using a "boing" sound from free YouTube audio library.

Just a proof-of-concept. Had to run npm i memfs to get it to work properly. Can use as template for more sounds, if desired - or discard if no sounds wanted at all.

Added "muted" member to global state so sound can be turned on and off - though nothing in the UI toggles it, yet.