Side-scroll wrap-around fix
Fixed up the problems with the previous change.

Since there are max 2 copies of the scene visible in canvas at once (when far left and far right edges touch and the boundary between them in on canvas) I've changed WrapSector enum to have only two values: Left and Right, designating those two copies of the scene.

By default, unless something needs to be shifted to paint itself inside the WrapSector.Right copy of the scene to be visible, it will paint itself in the WrapSector.Left copy of the scene (even if that's beyond the left edge of the canvas and therefore not actually visible).

There's also a Coord::toString method that I was using for debugging.

I've re-enabled collisions because they work properly now.

Previously there were 4 invisible colliders above, below, to the left, and to the right of the world. Now there are only two - one above the world, and one below. Left and right now wrap-around, so there shouldn't be colliders there.

There's also a black rectangle being drawn around the WrapSector.Left copy of the background image. That was for debugging. We can just remove it.