Narrow maps are not wrap-around
Updated logic so that maps more narrow than the canvas are treated as non-ring-world bounded map.

In conjunction with the above, the areas beyond the top and bottom edges of the background are not permitted to be visible on canvas. The same is true for the areas to the left and right of the background in the case of non-ring-world maps (except for a little padding to compensate for maps, like the current one, that are actually more narrow than the canvas).