NPCs come and go
If, after 5 seconds of being on screen, an NPC wanders off screen, the are in the "holding zone" (invisible = true and isOffScreen = true). They have a 3 in 10,000 chance of return each frame.

If you reappear on the left, your bias is to move right.

If you reappear on the right, your bias is to move left.