Slingshotting groundwork 1
Planets have canSlingshot and isSlingshotting members. "can" is specific to the type of planet. "is" is set randomly when an instance of that planet is spawned.

Global state has a slingshotAllowed boolean. Initially true, it can be set to false in advance of a black hole so that the black hole won't overlap with a slingshot event.

No actual slingshotting occurs yet. Remains to be implemented.